feat: change .npmrc path, default .npmrc file will write in `npm-folder`
new flag `npm-rc-user-home-enable` enable .npmrc file write user home
