+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "woodpecker-npm",
-  "version": "1.0.1",
+  "version": "1.1.0",
   "description": "publish package by npm for woodpecker-ci",
   "directories": {
     "doc": "doc"
